chrome/content/mochitest/tests/test_elemhide.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Element hiding tests</title>

  <link rel="stylesheet" type="text/css" href="/content/tests/SimpleTest/test.css" />

  <script type="text/javascript" src="/content/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/content/tests/SimpleTest/SimpleTest.js"></script>        

  <script type="application/x-javascript;version=1.7" src="common.js"></script>
  <script type="application/x-javascript;version=1.7" src="../httpd.js"></script>
</head>
<body>
  <p id="display"></p>
  <div id="content" style="visibility: hidden;">
    <iframe id="frame" onload="doRunTest()"></iframe>
  </div>

  <pre id="test">
  <script type="application/x-javascript;version=1.7">
    let abp = Components.classes["@mozilla.org/adblockplus;1"].createInstance().wrappedJSObject;

    let tests = [
      [[], "visible visible"],
      [["#div(test1)"], "hidden visible"],
      [["localhost#div(test1)"], "hidden visible"],
      [["localhost#div(test1)", "foo,foo2#p(test2)"], "hidden visible"],
      [["localhost,foo#div(test1)", "foo,localhost#p(test2)"], "hidden hidden"],
      [["localhost#div(test1)", "localhost#p(test2)"], "hidden hidden"],
      [["foo#div(test1)", "foo#p(test2)"], "visible visible"],

      [["localhost#div(testClass)"], "hidden visible"],
      [["localhost#p(testClass)"], "visible hidden"],
      [["localhost#*(testClass)"], "hidden hidden"],
      [["localhost#div(testClass)", "localhost#p(test2)"], "hidden hidden"],
      [["localhost#p(testClass)", "localhost#p(test2)"], "visible hidden"],
      [["localhost#p(testClass)(test2)"], "visible visible"],   // this filter is invalid, must be ignored

      [["localhost#*(id^=test)"], "hidden hidden"],
      [["localhost#p(id^=test)"], "visible hidden"],
      [["localhost#*(id$=2)"], "visible hidden"],
      [["localhost#p(id$=2)"], "visible hidden"],
      [["localhost#div(id$=2)"], "visible visible"],

      [["localhost#*(test1)(id^=test)"], "hidden visible"],
      [["localhost#*(testClass)(id^=test)"], "hidden hidden"],
      [["localhost#p(testClass)(id^=test)"], "visible hidden"],
      [["localhost#*(test1)(id$=2)"], "visible visible"],
      [["localhost#*(testClass)(id$=2)"], "visible hidden"],
      [["localhost#p(testClass)(id$=2)"], "visible hidden"],

      [["localhost#*(test1)(id^=test)(id$=2)"], "visible visible"],
      [["localhost#*(test1)(id^=test)(id$=1)"], "hidden visible"],
      [["localhost#p(test1)(id^=test)(id$=1)"], "visible visible"],
      [["localhost#div(test1)(id^=test)(id$=1)"], "hidden visible"],
      [["localhost#*(id^=test)(id$=2)"], "visible hidden"],
      [["localhost#*(id^=test)(id$=1)"], "hidden visible"],
      [["localhost#p(id^=test)(id$=1)"], "visible visible"],
      [["localhost#div(id^=test)(id$=1)"], "hidden visible"],

      [["localhost##div#test1"], "hidden visible"],
      [["localhost##p.testClass"], "visible hidden"],
      [["localhost##div#test1, p.testClass"], "hidden hidden"],
      [["localhost##div#test1", "localhost##p.testClass"], "hidden hidden"],
      [["localhost##.testClass"], "hidden hidden"],
    ];
    let currentTest = -1;
    let oldEnabled = abp.prefs.enabled;

    function runNextTest()
    {
      if (currentTest >= tests.length - 1)
      {
        abp.filterStorage.triggerSubscriptionObservers("reload", abp.filterStorage.subscriptions);
        abp.prefs.enabled = oldEnabled;
        SimpleTest.finish();
        return;
      }

      currentTest++;
      abp.elemhide.clear();
      document.getElementById("frame").contentWindow.location.href = "http://localhost:1234/test";
    }

    function doRunTest()
    {
      if (currentTest < 0)
        return;

      let [filters, expected] = tests[currentTest];
      for each (let filter in filters)
        abp.elemhide.add(abp.Filter.fromText(filter));
      abp.elemhide.apply();

      setTimeout(function()
      {
        let doc = document.getElementById("frame").contentDocument;
        let result = (doc.getElementById("test1").offsetHeight > 0 ? "visible" : "hidden") + " " + (doc.getElementById("test2").offsetHeight > 0 ? "visible" : "hidden");
        is(result, expected, filters.join("\n"));
        runNextTest();
      }, 0);
    }

    let server = new nsHttpServer();
    SimpleTest.waitForExplicitFinish();
    addLoadEvent(function()
    {
      server.start(1234);
      window.addEventListener("unload", function() { server.stop(); }, false);

      server.registerPathHandler("/test", function(metadata, response)
      {
        let body = '<div id="test1" class="testClass">foo</div><p id="test2" class="testClass">bar</p>';
        response.setStatusLine("1.1", "200", "OK");
        response.setHeader("Content-Type", "text/html");
        response.bodyOutputStream.write(body, body.length);
      });

      abp.prefs.enabled = true;
      runNextTest();
    });
  </script>
  </pre>
</body>
</html>
